fix: moved api folder into lib
