--- data/approve-tokens.mdx
@@ -1,3 +1,0 @@
-In order to stake your LPT you must first grant permissions to the Livepeer
-`BondingManager` contract. This will allow the contract to transfer the LPT you
-decide to stake, allowing you to start earning rewards.
